requirements.txt
Pillow
autopep8
